test: move mocking code
